magic_gradio_proxy/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio_client==0.7.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1